Refactor mockserv to enable unit testing
